feat: Rework debug functionalities for sunshine CSV
-  Move encrypt to same file as decrypt, provide a script to decrypt/encrypt the CSV data
- Add command to output JSON parsed from sunshine CSV
- Add documentation in Readme
- Add ability to output the data as the JSON used in the app (debugging)
